requirements.txt
Pillow
opencv-python
pdf2image
numpy
openai